requirements.txt
torch==1.10.0
torchvision==0.11.0
timm==0.6.0
pytorch-metric-learning==1.3.0
numpy==1.20.0
pillow==9.0.0
scikit-learn==1.0.0
tqdm==4.60.0
einops==0.4.0
matplotlib==3.5.0